Fix CI failures: Move type definitions out of server components to resolve Next.js 13+ App Router compatibility issues
