+++ src/types/index.ts
@@ -10,4 +10,4 @@ export const TaskSchema = z.object({
     .default(''),
 })
 export type Task = z.infer<typeof TaskSchema>
-export type { ResearchTaskInput } from '@/ai/flows/research-task'
+export type { ResearchTaskInput } from '@/types/research-task-types'
